agents/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai==0.8.3
pydantic==2.9.2
pydantic-settings==2.6.1
redis==5.2.0
httpx==0.27.2
python-dotenv==1.0.1
